requirements.txt
cffi==1.12.2
einky==0.0.4
numpy==1.16.2
peewee==3.8.2
Pillow==5.4.1
pkg-resources==0.0.0
psycopg2==2.7.7
pycparser==2.19
RPi.GPIO==0.6.5
smbus-cffi==0.5.1
spidev==3.4
